Refactor: Decouple CalendarView into multiple Swift files and fix drag flicker
- CalendarView.swift: Main calendar screen (shows month grid, toggles backlog).
- CalendarView+Logic.swift: All fetch methods, date-building utilities, and helper logic.
- MonthNavigationView.swift: Month navigation bar with <Prev / Next> buttons.
- DayCellView.swift: A single day cell with .onDrop for rescheduling tasks.
- BacklogListView.swift: Overdue task list with .onDrag.
- DayTasksListView.swift: Selected-day task list with .onDrag.

Fixes flicker by specifying a tiny transparent drag preview in DayTasksListView.
