+++ Views/BacklogListView.swift
@@ -1,29 +1,34 @@
-//// BacklogListView.swift
-//import SwiftUI
-//
-//struct BacklogListView: View {
-//    @Binding var backlogTasks: [TimeBox_Task]
-//    let deleteAction: (IndexSet) -> Void
-//
-//    var body: some View {
-//        if backlogTasks.isEmpty {
-//            Text("No overdue tasks!")
-//                .foregroundColor(.secondary)
-//                .padding()
-//            Spacer()
-//        } else {
-//            List {
-//                ForEach(backlogTasks, id: \.objectID) { task in
-//                    HStack {
-//                        Text(task.title ?? "Untitled")
-//                        Spacer()
-//                        Image(systemName: "exclamationmark.triangle.fill")
-//                            .foregroundColor(.orange)
-//                    }
-//                    .padding(.vertical, 6)
-//                }
-//                .onDelete(perform: deleteAction)
-//            }
-//        }
-//    }
-//}
+// BacklogListView.swift
+import SwiftUI
+
+struct BacklogListView: View {
+    @Binding var backlogTasks: [TimeBox_Task]
+    let deleteAction: (IndexSet) -> Void
+    
+    var body: some View {
+        if backlogTasks.isEmpty {
+            Text("No overdue tasks!")
+                .foregroundColor(.secondary)
+                .padding()
+            Spacer()
+        } else {
+            List {
+                ForEach(backlogTasks, id: \.objectID) { task in
+                    HStack {
+                        Text(task.title ?? "Untitled")
+                        Spacer()
+                        Image(systemName: "exclamationmark.triangle.fill")
+                            .foregroundColor(.orange)
+                    }
+                    .padding(.vertical, 6)
+                    // DRAG
+                    .onDrag {
+                        let taskID = task.objectID.uriRepresentation().absoluteString
+                        return NSItemProvider(object: taskID as NSString)
+                    }
+                }
+                .onDelete(perform: deleteAction)
+            }
+        }
+    }
+}
