+++ Views/CalendarView+Logic.swift .swift
@@ -5,7 +5,6 @@ import CoreData
 
 extension CalendarView {
     
-    // Build an array of days in the given month
     func makeDaysInMonth(_ date: Date) -> [Date] {
         guard let interval = Calendar.current.dateInterval(of: .month, for: date) else { return [] }
         var days: [Date] = []
@@ -15,12 +14,6 @@ extension CalendarView {
             current = Calendar.current.date(byAdding: .day, value: 1, to: current) ?? current
         }
         return days
-    }
-    
-    func formatMonth(_ date: Date) -> String {
-        let formatter = DateFormatter()
-        formatter.dateFormat = "LLLL yyyy"
-        return formatter.string(from: date)
     }
     
     func fetchTasks(for day: Date) -> [TimeBox_Task] {
@@ -33,13 +26,13 @@ extension CalendarView {
         let request = NSFetchRequest<TimeBox_Task>(entityName: "TimeBox_Task")
         request.predicate = NSPredicate(
             format: "startTime >= %@ AND startTime < %@",
-            startOfDay as CVarArg,
-            endOfDay as CVarArg
+            startOfDay as CVarArg, endOfDay as CVarArg
         )
+        
         do {
             return try viewContext.fetch(request)
         } catch {
-            print("DEBUG: Error fetching tasks for day:", error)
+            print("DEBUG: Error fetching tasks for \(day):", error)
             return []
         }
     }
@@ -50,16 +43,15 @@ extension CalendarView {
         let request = NSFetchRequest<TimeBox_Task>(entityName: "TimeBox_Task")
         request.predicate = NSPredicate(
             format: "startTime >= %@ AND startTime < %@",
-            interval.start as CVarArg,
-            interval.end as CVarArg
+            interval.start as CVarArg, interval.end as CVarArg
         )
         do {
             let tasks = try viewContext.fetch(request)
             var counts: [Date: Int] = [:]
             for t in tasks {
                 if let st = t.startTime {
-                    let dayOnly = Calendar.current.startOfDay(for: st)
-                    counts[dayOnly, default: 0] += 1
+                    let d = Calendar.current.startOfDay(for: st)
+                    counts[d, default: 0] += 1
                 }
             }
             dailyTaskCounts = counts
@@ -71,6 +63,7 @@ extension CalendarView {
     
     func fetchBacklogTasks() -> [TimeBox_Task] {
         let todayStart = Calendar.current.startOfDay(for: Date())
+        
         let request = NSFetchRequest<TimeBox_Task>(entityName: "TimeBox_Task")
         request.predicate = NSPredicate(
             format: "startTime < %@ AND status != %@",
