+++ Views/MonthNavigationView.swift
@@ -1,28 +1,35 @@
 // MonthNavigationView.swift
+
 import SwiftUI
 
 struct MonthNavigationView: View {
-    @Binding var currentMonth: Date
-    let fetchMonthlyTaskCounts: (Date) -> Void
-
+    let currentMonth: Date
+    let onPrev: () -> Void
+    let onNext: () -> Void
+    
+    // Whether we're in backlog mode (to possibly hide month text)
+    let showBacklog: Bool
+    
     var body: some View {
         HStack {
-            Button("< Prev") {
-                currentMonth = Calendar.current.date(byAdding: .month, value: -1, to: currentMonth) ?? currentMonth
-                fetchMonthlyTaskCounts(currentMonth)
-            }
+            Button("< Prev") { onPrev() }
             Spacer()
-            Text(formatMonth(currentMonth))
-                .font(.headline)
-            Spacer()
-            Button("Next >") {
-                currentMonth = Calendar.current.date(byAdding: .month, value: 1, to: currentMonth) ?? currentMonth
-                fetchMonthlyTaskCounts(currentMonth)
+            
+            if !showBacklog {
+                Text(formatMonth(currentMonth))
+                    .font(.headline)
+            } else {
+                Text("")  // hide text if backlog
+                    .font(.headline)
             }
+            
+            Spacer()
+            Button("Next >") { onNext() }
         }
         .padding()
     }
-
+    
+    // Local helper for formatting the month
     private func formatMonth(_ date: Date) -> String {
         let formatter = DateFormatter()
         formatter.dateFormat = "LLLL yyyy"
